LayoutTests/editing/undomanager/automatic-transaction-data.html

<!DOCTYPE html>
<html>
<head>
<title>automatic transaction on node data</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
This tests that UndoManager correctly reverts/reapplies automatic DOM transactions
on node data.

<div id="div"></div>
<script>
var undoManager = document.undoManager;
var div = document.getElementById("div");

test(function() {
	div.innerText = "123456789";
    undoManager.transact({
        "executeAutomatic": function() { div.firstChild.replaceData(3, 3, "abcd"); }
    });
    assert_equals(div.innerText, "123abcd789");
}, "Started with 123456789. Transacted replaceData(3, 3, 'abcd'). Got 123abcd789");

test(function() {
    undoManager.undo();
    assert_equals(div.innerText, "123456789");
}, "undoManager.undo(). Got 123456789");

test(function() {
    undoManager.redo();
    assert_equals(div.innerText, "123abcd789");
}, "undoManager.redo(). Got 123abcd789");

test(function() {
	div.innerText = "1234";
    undoManager.transact({
        "executeAutomatic": function() { div.firstChild.replaceData(2, 10, "a"); }
    });
    assert_equals(div.innerText, "12a");
}, "Started with 1234. Transacted replaceData(2, 10, 'a'). Got 12a");

test(function() {
    undoManager.undo();
    assert_equals(div.innerText, "1234");
}, "undoManager.undo(). Got 1234");

test(function() {
    undoManager.redo();
    assert_equals(div.innerText, "12a");
}, "undoManager.redo(). Got 12a");

div.style.display = "none";
</script>
</body>
</html>
